Added some logging.  Simplify roundcount parser.

+++ src/AmmoFinder.Common/Models/ProductModel.cs
@@ -8,7 +8,7 @@ namespace AmmoFinder.Common.Models
         public string Description { get; set; }
         public string Brand { get; set; }
         public bool IsAvailable { get; set; }
-        public int Inventory { get; set; }
+        public long Inventory { get; set; }
         public decimal Price { get; set; }
         public int RoundCount { get; set; }
         public string RoundType { get; set; }
